+++ CHANGELOG.md
@@ -6,6 +6,8 @@ read by developers.
 
 ### Added
 - Add inline destinations to user edit page in admin
+- Add management command for listing filters
+- Add management command for bulk acting on incidents matching a given filter
 
 ### Fixed
 - Raise validation error on posting incident with tags without tag key